% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Anchors}
\alias{Anchors}
\alias{END}
\alias{START}
\title{The start or end of a string.}
\format{\preformatted{Classes 'regex', 'character'  chr "^"
}}
\usage{
START

END
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Match the start or end of a string.
}
\examples{
START
END
}
\references{
\url{http://www.regular-expressions.info/anchors.html}
}
\keyword{datasets}

