% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regex-methods.R
\name{as.regex}
\alias{as.perl_regex}
\alias{as.regex}
\alias{is.perl_regex}
\alias{is.regex}
\title{Convert or test for regex objects}
\usage{
as.regex(x)

is.regex(x)

as.perl_regex(x)

is.perl_regex(x)
}
\arguments{
\item{x}{An object to test or convert.}
}
\value{
\code{as.regex} returns the inputs object, with class
\code{c("regex", "character")}.
\code{is.regex} returns \code{TRUE} when the input inherits from class
\code{"regex"} and \code{FALSE} otherwise.
Similarly \code{as.perl_regex} gives the input object a class of
\code{c("perl_regex", "regex", "character")}.  For compatibility with
the \code{stringr} package it also sets the attribute \code{perl = TRUE}.
}
\description{
\code{as.regex} gives objects the class \code{"regex"}. \code{is.regex}
tests for objects of class \code{"regex"}.
}
\examples{
x <- as.regex("month.abb")
is.regex(x)
as.perl_regex("month.abb")
is.perl_regex(lookahead("a"))
}

