% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-base.R
\name{modify_mode}
\alias{case_insensitive}
\alias{duplicate_group_names}
\alias{free_spacing}
\alias{modify_mode}
\alias{multi_line}
\alias{no_backslash_escaping}
\alias{single_line}
\title{Apply mode modifiers}
\description{
See \code{\link[rebus.base]{modify_mode}}.
}

