% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backreferences.R
\name{recursive}
\alias{recursive}
\title{Make the regular expression recursive.}
\usage{
recursive(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Makes the regular expression (or part of it) recursive.
}
\examples{
recursive("a")
}
\references{
\url{http://www.regular-expressions.info/recurse.html}
}

