% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-unicode.R
\name{UnicodeGeneralCategory}
\alias{UGC_CASED_LETTER}
\alias{UGC_CLOSE_PUNCTUATION}
\alias{UGC_CONNECTOR_PUNCTUATION}
\alias{UGC_CONTROL}
\alias{UGC_CURRENCY_SYMBOL}
\alias{UGC_DASH_PUNCTUATION}
\alias{UGC_DECIMAL_NUMBER}
\alias{UGC_ENCLOSING_MARK}
\alias{UGC_FINAL_PUNCTUATION}
\alias{UGC_FORMAT_CONTROL}
\alias{UGC_INITIAL_PUNCTUATION}
\alias{UGC_LETTER}
\alias{UGC_LETTER_NUMBER}
\alias{UGC_LINE_SEPARATOR}
\alias{UGC_LOWERCASE_LETTER}
\alias{UGC_MARK}
\alias{UGC_MATH_SYMBOL}
\alias{UGC_MODIFIER_LETTER}
\alias{UGC_MODIFIER_SYMBOL}
\alias{UGC_NONSPACING_MARK}
\alias{UGC_NUMBER}
\alias{UGC_OPEN_PUNCTUATION}
\alias{UGC_OTHER}
\alias{UGC_OTHER_LETTER}
\alias{UGC_OTHER_NUMBER}
\alias{UGC_OTHER_PUNCTUATION}
\alias{UGC_OTHER_SYMBOL}
\alias{UGC_PARAGRAPH_SEPARATOR}
\alias{UGC_PRIVATE_USE_CONTROL}
\alias{UGC_PUNCTUATION}
\alias{UGC_SEPARATOR}
\alias{UGC_SPACE_SEPARATOR}
\alias{UGC_SPACING_MARK}
\alias{UGC_SURROGATE_CONTROL}
\alias{UGC_SYMBOL}
\alias{UGC_TITLECASE_LETTER}
\alias{UGC_UNASSIGNED_CONTROL}
\alias{UGC_UPPERCASE_LETTER}
\alias{UnicodeGeneralCategory}
\alias{ugc_cased_letter}
\alias{ugc_close_punctuation}
\alias{ugc_connector_punctuation}
\alias{ugc_control}
\alias{ugc_currency_symbol}
\alias{ugc_dash_punctuation}
\alias{ugc_decimal_number}
\alias{ugc_enclosing_mark}
\alias{ugc_final_punctuation}
\alias{ugc_format_control}
\alias{ugc_initial_punctuation}
\alias{ugc_letter}
\alias{ugc_letter_number}
\alias{ugc_line_separator}
\alias{ugc_lowercase_letter}
\alias{ugc_mark}
\alias{ugc_math_symbol}
\alias{ugc_modifier_letter}
\alias{ugc_modifier_symbol}
\alias{ugc_nonspacing_mark}
\alias{ugc_number}
\alias{ugc_open_punctuation}
\alias{ugc_other}
\alias{ugc_other_letter}
\alias{ugc_other_number}
\alias{ugc_other_punctuation}
\alias{ugc_other_symbol}
\alias{ugc_paragraph_separator}
\alias{ugc_private_use_control}
\alias{ugc_punctuation}
\alias{ugc_separator}
\alias{ugc_space_separator}
\alias{ugc_spacing_mark}
\alias{ugc_surrogate_control}
\alias{ugc_symbol}
\alias{ugc_titlecase_letter}
\alias{ugc_unassigned_control}
\alias{ugc_uppercase_letter}
\title{Unicode General Categories}
\description{
See \code{\link[rebus.unicode]{UnicodeGeneralCategory}}.
}

