% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-unicode.R
\name{UnicodeProperty}
\alias{UnicodeProperty}
\alias{up_alphabetic}
\alias{UP_ALPHABETIC}
\alias{up_ascii_hex_digit}
\alias{UP_ASCII_HEX_DIGIT}
\alias{up_bidi_control}
\alias{UP_BIDI_CONTROL}
\alias{up_bidi_mirrored}
\alias{UP_BIDI_MIRRORED}
\alias{up_case_ignorable}
\alias{UP_CASE_IGNORABLE}
\alias{up_case_sensitive}
\alias{UP_CASE_SENSITIVE}
\alias{up_cased}
\alias{UP_CASED}
\alias{up_changes_when_casefolded}
\alias{UP_CHANGES_WHEN_CASEFOLDED}
\alias{up_changes_when_casemapped}
\alias{UP_CHANGES_WHEN_CASEMAPPED}
\alias{up_changes_when_lowercased}
\alias{UP_CHANGES_WHEN_LOWERCASED}
\alias{up_changes_when_nfkc_casefolded}
\alias{UP_CHANGES_WHEN_NFKC_CASEFOLDED}
\alias{up_changes_when_titlecased}
\alias{UP_CHANGES_WHEN_TITLECASED}
\alias{up_changes_when_uppercased}
\alias{UP_CHANGES_WHEN_UPPERCASED}
\alias{up_dash}
\alias{UP_DASH}
\alias{up_default_ignorable_code_point}
\alias{UP_DEFAULT_IGNORABLE_CODE_POINT}
\alias{up_deprecated}
\alias{UP_DEPRECATED}
\alias{up_diacritic}
\alias{UP_DIACRITIC}
\alias{up_extender}
\alias{UP_EXTENDER}
\alias{up_hex_digit}
\alias{UP_HEX_DIGIT}
\alias{up_hyphen}
\alias{UP_HYPHEN}
\alias{up_id_continue}
\alias{UP_ID_CONTINUE}
\alias{up_id_start}
\alias{UP_ID_START}
\alias{up_ideographic}
\alias{UP_IDEOGRAPHIC}
\alias{up_lowercase}
\alias{UP_LOWERCASE}
\alias{up_math}
\alias{UP_MATH}
\alias{up_noncharacter_code_point}
\alias{UP_NONCHARACTER_CODE_POINT}
\alias{up_posix_alnum}
\alias{UP_POSIX_ALNUM}
\alias{up_posix_blank}
\alias{UP_POSIX_BLANK}
\alias{up_posix_graph}
\alias{UP_POSIX_GRAPH}
\alias{up_posix_print}
\alias{UP_POSIX_PRINT}
\alias{up_posix_xdigit}
\alias{UP_POSIX_XDIGIT}
\alias{up_quotation_mark}
\alias{UP_QUOTATION_MARK}
\alias{up_soft_dotted}
\alias{UP_SOFT_DOTTED}
\alias{up_terminal_punctuation}
\alias{UP_TERMINAL_PUNCTUATION}
\alias{up_uppercase}
\alias{UP_UPPERCASE}
\alias{up_white_space}
\alias{UP_WHITE_SPACE}
\title{Unicode Properties}
\description{
See \code{\link[rebus.unicode]{UnicodeProperty}}.
}
