% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WordBoundaries}
\alias{BOUNDARY}
\alias{NOT_BOUNDARY}
\alias{WordBoundaries}
\title{Word boundaries}
\format{\preformatted{Classes 'regex', 'character'  chr "\\\\b"
}}
\usage{
BOUNDARY

NOT_BOUNDARY
}
\description{
Match a word boundary.
}
\examples{
BOUNDARY
NOT_BOUNDARY
}
\references{
\url{http://www.regular-expressions.info/wordboundaries.html}
}
\seealso{
\code{\link{ALPHA}}, \code{\link{BACKSLASH}}, \code{\link{START}}
}
\keyword{datasets}

