% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-base.R
\name{CharacterClasses}
\alias{CharacterClasses}
\alias{ALNUM}
\alias{ALPHA}
\alias{BLANK}
\alias{CNTRL}
\alias{DIGIT}
\alias{GRAPH}
\alias{LOWER}
\alias{PRINT}
\alias{PUNCT}
\alias{SPACE}
\alias{UPPER}
\alias{HEX_DIGIT}
\alias{ANY_CHAR}
\alias{GRAPHEME}
\alias{NEWLINE}
\alias{DGT}
\alias{WRD}
\alias{SPC}
\alias{NOT_DGT}
\alias{NOT_WRD}
\alias{NOT_SPC}
\alias{ASCII_DIGIT}
\alias{ASCII_LOWER}
\alias{ASCII_UPPER}
\alias{ASCII_ALPHA}
\alias{ASCII_ALNUM}
\alias{UNMATCHABLE}
\title{Class Constants}
\description{
See \code{\link[rebus.base]{CharacterClasses}}.
}
