% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-base.R
\name{ClassGroups}
\alias{ClassGroups}
\alias{alnum}
\alias{alpha}
\alias{any_char}
\alias{ascii_alnum}
\alias{ascii_alpha}
\alias{ascii_digit}
\alias{ascii_lower}
\alias{ascii_upper}
\alias{blank}
\alias{char_range}
\alias{cntrl}
\alias{dgt}
\alias{digit}
\alias{graph}
\alias{hex_digit}
\alias{lower}
\alias{not_dgt}
\alias{not_spc}
\alias{not_wrd}
\alias{printable}
\alias{punct}
\alias{space}
\alias{spc}
\alias{upper}
\alias{wrd}
\title{Character classes}
\description{
See \code{\link[rebus.base]{ClassGroups}}.
}

