% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-base.R
\name{SpecialCharacters}
\alias{BACKSLASH}
\alias{CARET}
\alias{CLOSE_BRACKET}
\alias{CLOSE_PAREN}
\alias{DOLLAR}
\alias{DOT}
\alias{OPEN_BRACE}
\alias{OPEN_BRACKET}
\alias{OPEN_PAREN}
\alias{PIPE}
\alias{PLUS}
\alias{QUESTION}
\alias{STAR}
\alias{SpecialCharacters}
\title{Special characters}
\description{
See \code{\link[rebus.base]{SpecialCharacters}}.
}

