% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SpecialCharacters}
\alias{BACKSLASH}
\alias{CARET}
\alias{CLOSE_PAREN}
\alias{DOLLAR}
\alias{DOT}
\alias{OPEN_BRACE}
\alias{OPEN_BRACKET}
\alias{OPEN_PAREN}
\alias{PIPE}
\alias{PLUS}
\alias{QUESTION}
\alias{STAR}
\alias{SpecialCharacters}
\title{Special characters}
\format{\preformatted{Classes 'regex', 'character'  chr "\\\\\\\\"
}}
\usage{
BACKSLASH

CARET

DOLLAR

DOT

PIPE

QUESTION

STAR

PLUS

OPEN_PAREN

CLOSE_PAREN

OPEN_BRACKET

OPEN_BRACE
}
\description{
Constants to match special characters
}
\examples{
BACKSLASH
CARET
DOLLAR
DOT
PIPE
QUESTION
STAR
PLUS
OPEN_PAREN
CLOSE_PAREN
OPEN_BRACKET
OPEN_BRACE
}
\references{
\url{http://www.regular-expressions.info/characters.html}
}
\seealso{
\code{\link{ALPHA}}, \code{\link{START}}, \code{\link{BOUNDARY}}
}
\keyword{datasets}

