% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-unicode.R
\name{Unicode}
\alias{Unicode}
\alias{ADDITIONAL_ARROWS}
\alias{AEGEAN_NUMBERS}
\alias{ALCHEMICAL_SYMBOLS}
\alias{ALPHABETIC_PRESENTATION_FORMS}
\alias{ANCIENT_GREEK_MUSICAL_NOTATION}
\alias{ANCIENT_GREEK_NUMBERS}
\alias{ANCIENT_SYMBOLS}
\alias{ARABIC}
\alias{ARABIC_EXTENDED_A}
\alias{ARABIC_MATHEMATICAL_ALPHANUMERIC_SYMBOLS}
\alias{ARABIC_PRESENTATION_FORMS_A}
\alias{ARABIC_PRESENTATION_FORMS_B}
\alias{ARABIC_SUPPLEMENT}
\alias{ARMENIAN}
\alias{ARMENIAN_LIGATURES}
\alias{AVESTAN}
\alias{BALINESE}
\alias{BAMUN}
\alias{BAMUN_SUPPLEMENT}
\alias{BASSA_VAH}
\alias{BATAK}
\alias{BENGALI_AND_ASSAMESE}
\alias{BLOCK_ELEMENTS}
\alias{BOPOMOFO}
\alias{BOPOMOFO_EXTENDED}
\alias{BOX_DRAWING}
\alias{BRAHMI}
\alias{BRAILLE_PATTERNS}
\alias{BUGINESE}
\alias{BUHID}
\alias{BYZANTINE_MUSICAL_SYMBOLS}
\alias{CARD_SUITS}
\alias{CARIAN}
\alias{CAUCASIAN_ALBANIAN}
\alias{CHAKMA}
\alias{CHAM}
\alias{CHEROKEE}
\alias{CHESS_CHECKERS_DRAUGHTS}
\alias{CJK_COMPATIBILITY}
\alias{CJK_COMPATIBILITY_FORMS}
\alias{CJK_COMPATIBILITY_IDEOGRAPHS}
\alias{CJK_COMPATIBILITY_IDEOGRAPHS_SUPPLEMENT}
\alias{CJK_IDEOGRAPHIC_DESCRIPTION_CHARACTERS}
\alias{CJK_STROKES}
\alias{CJK_SYMBOLS_AND_PUNCTUATION}
\alias{CJK_UNIFIED_IDEOGRAPHS}
\alias{CJK_UNIFIED_IDEOGRAPHS_EXTENSION_A}
\alias{CJK_UNIFIED_IDEOGRAPHS_EXTENSION_B}
\alias{CJK_UNIFIED_IDEOGRAPHS_EXTENSION_C}
\alias{CJK_UNIFIED_IDEOGRAPHS_EXTENSION_D}
\alias{COMBINING_DIACRITIC_EXTENDED}
\alias{COMBINING_DIACRITIC_MARKS}
\alias{COMBINING_DIACRITIC_MARKS_FOR_SYMBOLS}
\alias{COMBINING_DIACRITIC_SUPPLEMENT}
\alias{COMBINING_HALF_MARKS}
\alias{COMMON_INDIC_NUMBER_FORMS}
\alias{CONTROL_PICTURES}
\alias{COPTIC}
\alias{COPTIC_EPACT_NUMBERS}
\alias{COUNTING_ROD_NUMERALS}
\alias{CUNEIFORM}
\alias{CUNEIFORM_NUMBERS_AND_PUNCTUATION}
\alias{CURRENCY_SYMBOLS}
\alias{CYPRIOT_SYLLABARY}
\alias{CYRILLIC}
\alias{CYRILLIC_EXTENDED_A}
\alias{CYRILLIC_EXTENDED_B}
\alias{CYRILLIC_SUPPLEMENT}
\alias{DESERET}
\alias{DEVANAGARI}
\alias{DEVANAGARI_EXTENDED}
\alias{DINGBATS}
\alias{DOMINO_TILES}
\alias{DUPLOYAN}
\alias{EGYPTIAN_HIEROGLYPHS}
\alias{ELBASAN}
\alias{EMOTICONS}
\alias{ENCLOSED_ALPHANUMERIC_SUPPLEMENT}
\alias{ENCLOSED_ALPHANUMERICS}
\alias{ENCLOSED_CJK_LETTERS_AND_MONTHS}
\alias{ENCLOSED_IDEOGRAPHIC_SUPPLEMENT}
\alias{ETHIOPIC}
\alias{ETHIOPIC_EXTENDED}
\alias{ETHIOPIC_EXTENDED_A}
\alias{ETHIOPIC_SUPPLEMENT}
\alias{FLOORS_AND_CEILINGS}
\alias{FULLWIDTH_ASCII_DIGITS}
\alias{FULLWIDTH_ASCII_PUNCTUATION}
\alias{GENERAL_PUNCTUATION}
\alias{GEOMETRIC_SHAPES}
\alias{GEOMETRIC_SHAPES_EXTENDED}
\alias{GEORGIAN}
\alias{GEORGIAN_SUPPLEMENT}
\alias{GLAGOLITIC}
\alias{GOTHIC}
\alias{GRANTHA}
\alias{GREEK_AND_COPTIC}
\alias{GREEK_EXTENDED}
\alias{GUJARATI}
\alias{GURMUKHI}
\alias{HALFWIDTH_AND_FULLWIDTH_FORMS}
\alias{HANGUL_COMPATIBILITY_JAMO}
\alias{HANGUL_JAMO}
\alias{HANGUL_JAMO_EXTENDED_A}
\alias{HANGUL_JAMO_EXTENDED_B}
\alias{HANGUL_SYLLABLES}
\alias{HANUNOO}
\alias{HEBREW}
\alias{HIRAGANA}
\alias{IMPERIAL_ARAMAIC}
\alias{INVISIBLE_OPERATORS}
\alias{IPA_EXTENSIONS}
\alias{JAPANESE_CHESS}
\alias{JAVANESE}
\alias{KAITHI}
\alias{KANA_SUPPLEMENT}
\alias{KANBUN}
\alias{KANGXI_RADICALS}
\alias{KANGXI_RADICALS_SUPPLEMENT}
\alias{KANNADA}
\alias{KATAKANA}
\alias{KATAKANA_PHONETIC_EXTENSIONS}
\alias{KAYAH_LI}
\alias{KHAROSHTHI}
\alias{KHMER}
\alias{KHMER_SYMBOLS}
\alias{KHOJKI}
\alias{KHUDAWADI}
\alias{LAO}
\alias{LATIN}
\alias{LATIN_1_PUNCTUATION}
\alias{LATIN_1_SUPPLEMENT}
\alias{LATIN_EXTENDED_A}
\alias{LATIN_EXTENDED_ADDITIONAL}
\alias{LATIN_EXTENDED_B}
\alias{LATIN_EXTENDED_C}
\alias{LATIN_EXTENDED_D}
\alias{LATIN_EXTENDED_E}
\alias{LATIN_LIGATURES}
\alias{LEPCHA}
\alias{LETTERLIKE_SYMBOLS}
\alias{LIMBU}
\alias{LINEAR_A}
\alias{LINEAR_B_IDEOGRAMS}
\alias{LINEAR_B_SYLLABARY}
\alias{LISU}
\alias{LYCIAN}
\alias{LYDIAN}
\alias{MAHAJANI}
\alias{MAHJONG_TILES}
\alias{MALAYALAM}
\alias{MANDAIC}
\alias{MANICHAEAN}
\alias{MATH_ARROWS}
\alias{MATHEMATICAL_ALPHANUMERIC_SYMBOLS}
\alias{MEETEI_MAYEK}
\alias{MEETEI_MAYEK_EXTENSIONS}
\alias{MENDE_KIKAKUI}
\alias{MEROITIC_CURSIVE}
\alias{MEROITIC_HIEROGLYPHS}
\alias{MIAO}
\alias{MISCELLANEOUS_MATHEMATICAL_SYMBOLS_A}
\alias{MISCELLANEOUS_MATHEMATICAL_SYMBOLS_B}
\alias{MISCELLANEOUS_SYMBOLS_AND_PICTOGRAPHS}
\alias{MISCELLANEOUS_TECHNICAL}
\alias{MODI}
\alias{MODIFIER_TONE_LETTERS}
\alias{MONGOLIAN}
\alias{MRO}
\alias{MUSICAL_SYMBOLS}
\alias{MYANMAR}
\alias{MYANMAR_EXTENDED_A}
\alias{MYANMAR_EXTENDED_B}
\alias{NABATAEAN}
\alias{NEW_TAI_LUE}
\alias{NKO}
\alias{NUMBER_FORMS}
\alias{OGHAM}
\alias{OL_CHIKI}
\alias{OLD_ITALIC}
\alias{OLD_NORTH_ARABIAN}
\alias{OLD_PERMIC}
\alias{OLD_PERSIAN}
\alias{OLD_SOUTH_ARABIAN}
\alias{OLD_TURKIC}
\alias{OPTICAL_CHARACTER_RECOGNITION}
\alias{ORIYA}
\alias{ORNAMENTAL_DINGBATS}
\alias{OSMANYA}
\alias{PAHAWH_HMONG}
\alias{PAHLAVI_INSCRIPTIONAL}
\alias{PAHLAVI_PSALTER}
\alias{PALMYRENE}
\alias{PAU_CIN_HAU}
\alias{PHAGS_PA}
\alias{PHAISTOS_DISC}
\alias{PHOENICIAN}
\alias{PHONETIC_EXTENSIONS}
\alias{PHONETIC_EXTENSIONS_SUPPLEMENT}
\alias{PLAYING_CARDS}
\alias{PRIVATE_USE_AREA}
\alias{REJANG}
\alias{RUMI_NUMERAL_SYMBOLS}
\alias{RUNIC}
\alias{SAMARITAN}
\alias{SAURASHTRA}
\alias{SHARADA}
\alias{SHAVIAN}
\alias{SHORTHAND_FORMAT_CONTROLS}
\alias{SIDDHAM}
\alias{SINHALA}
\alias{SINHALA_ARCHAIC_NUMBERS}
\alias{SMALL_FORM_VARIANTS}
\alias{SORA_SOMPENG}
\alias{SPACING_MODIFIER_LETTERS}
\alias{SPECIALS}
\alias{SUNDANESE}
\alias{SUNDANESE_SUPPLEMENT}
\alias{SUPERSCRIPTS_AND_SUBSCRIPTS}
\alias{SUPPLEMENTAL_ARROWS_A}
\alias{SUPPLEMENTAL_MATHEMATICAL_OPERATORS}
\alias{SUPPLEMENTAL_PUNCTUATION}
\alias{SUPPLEMENTARY_PRIVATE_USE_AREA_A}
\alias{SUPPLEMENTARY_PRIVATE_USE_AREA_B}
\alias{SYLOTI_NAGRI}
\alias{SYRIAC}
\alias{TAGALOG}
\alias{TAGBANWA}
\alias{TAGS}
\alias{TAI_LE}
\alias{TAI_THAM}
\alias{TAI_VIET}
\alias{TAI_XUAN_JING_SYMBOLS}
\alias{TAKRI}
\alias{TAMIL}
\alias{TELUGU}
\alias{THAANA}
\alias{THAI}
\alias{TIBETAN}
\alias{TIFINAGH}
\alias{TIRHUTA}
\alias{TRANSPORT_AND_MAP_SYMBOLS}
\alias{UGARITIC}
\alias{UNIFIED_CANADIAN_ABORIGINAL_SYLLABICS}
\alias{UNIFIED_CANADIAN_ABORIGINAL_SYLLABICS_EXTENDED}
\alias{VAI}
\alias{VARIATION_SELECTORS}
\alias{VARIATION_SELECTORS_SUPPLEMENT}
\alias{VEDIC_EXTENSIONS}
\alias{VERTICAL_FORMS}
\alias{WARANG_CITI}
\alias{YI_RADICALS}
\alias{YI_SYLLABLES}
\alias{YIJING_HEXAGRAM_SYMBOLS}
\alias{YIJING_MONO_DI_AND_TRIGRAMS}
\alias{additional_arrows}
\alias{aegean_numbers}
\alias{alchemical_symbols}
\alias{alphabetic_presentation_forms}
\alias{ancient_greek_musical_notation}
\alias{ancient_greek_numbers}
\alias{ancient_symbols}
\alias{arabic}
\alias{arabic_extended_a}
\alias{arabic_mathematical_alphanumeric_symbols}
\alias{arabic_presentation_forms_a}
\alias{arabic_presentation_forms_b}
\alias{arabic_supplement}
\alias{armenian}
\alias{armenian_ligatures}
\alias{avestan}
\alias{balinese}
\alias{bamun}
\alias{bamun_supplement}
\alias{bassa_vah}
\alias{batak}
\alias{bengali_and_assamese}
\alias{block_elements}
\alias{bopomofo}
\alias{bopomofo_extended}
\alias{box_drawing}
\alias{brahmi}
\alias{braille_patterns}
\alias{buginese}
\alias{buhid}
\alias{byzantine_musical_symbols}
\alias{card_suits}
\alias{carian}
\alias{caucasian_albanian}
\alias{chakma}
\alias{cham}
\alias{cherokee}
\alias{chess_checkers_draughts}
\alias{cjk_compatibility}
\alias{cjk_compatibility_forms}
\alias{cjk_compatibility_ideographs}
\alias{cjk_compatibility_ideographs_supplement}
\alias{cjk_ideographic_description_characters}
\alias{cjk_strokes}
\alias{cjk_symbols_and_punctuation}
\alias{cjk_unified_ideographs}
\alias{cjk_unified_ideographs_extension_a}
\alias{cjk_unified_ideographs_extension_b}
\alias{cjk_unified_ideographs_extension_c}
\alias{cjk_unified_ideographs_extension_d}
\alias{combining_diacritic_extended}
\alias{combining_diacritic_marks}
\alias{combining_diacritic_marks_for_symbols}
\alias{combining_diacritic_supplement}
\alias{combining_half_marks}
\alias{common_indic_number_forms}
\alias{control_pictures}
\alias{coptic}
\alias{coptic_epact_numbers}
\alias{counting_rod_numerals}
\alias{cuneiform}
\alias{cuneiform_numbers_and_punctuation}
\alias{currency_symbols}
\alias{cypriot_syllabary}
\alias{cyrillic}
\alias{cyrillic_extended_a}
\alias{cyrillic_extended_b}
\alias{cyrillic_supplement}
\alias{deseret}
\alias{devanagari}
\alias{devanagari_extended}
\alias{dingbats}
\alias{domino_tiles}
\alias{duployan}
\alias{egyptian_hieroglyphs}
\alias{elbasan}
\alias{emoticons}
\alias{enclosed_alphanumeric_supplement}
\alias{enclosed_alphanumerics}
\alias{enclosed_cjk_letters_and_months}
\alias{enclosed_ideographic_supplement}
\alias{ethiopic}
\alias{ethiopic_extended}
\alias{ethiopic_extended_a}
\alias{ethiopic_supplement}
\alias{floors_and_ceilings}
\alias{fullwidth_ascii_digits}
\alias{fullwidth_ascii_punctuation}
\alias{general_punctuation}
\alias{geometric_shapes}
\alias{geometric_shapes_extended}
\alias{georgian}
\alias{georgian_supplement}
\alias{glagolitic}
\alias{gothic}
\alias{grantha}
\alias{greek_and_coptic}
\alias{greek_extended}
\alias{gujarati}
\alias{gurmukhi}
\alias{halfwidth_and_fullwidth_forms}
\alias{hangul_compatibility_jamo}
\alias{hangul_jamo}
\alias{hangul_jamo_extended_a}
\alias{hangul_jamo_extended_b}
\alias{hangul_syllables}
\alias{hanunoo}
\alias{hebrew}
\alias{hiragana}
\alias{imperial_aramaic}
\alias{invisible_operators}
\alias{ipa_extensions}
\alias{japanese_chess}
\alias{javanese}
\alias{kaithi}
\alias{kana_supplement}
\alias{kanbun}
\alias{kangxi_radicals}
\alias{kangxi_radicals_supplement}
\alias{kannada}
\alias{katakana}
\alias{katakana_phonetic_extensions}
\alias{kayah_li}
\alias{kharoshthi}
\alias{khmer}
\alias{khmer_symbols}
\alias{khojki}
\alias{khudawadi}
\alias{lao}
\alias{latin}
\alias{latin_1_punctuation}
\alias{latin_1_supplement}
\alias{latin_extended_a}
\alias{latin_extended_additional}
\alias{latin_extended_b}
\alias{latin_extended_c}
\alias{latin_extended_d}
\alias{latin_extended_e}
\alias{latin_ligatures}
\alias{lepcha}
\alias{letterlike_symbols}
\alias{limbu}
\alias{linear_a}
\alias{linear_b_ideograms}
\alias{linear_b_syllabary}
\alias{lisu}
\alias{lycian}
\alias{lydian}
\alias{mahajani}
\alias{mahjong_tiles}
\alias{malayalam}
\alias{mandaic}
\alias{manichaean}
\alias{math_arrows}
\alias{mathematical_alphanumeric_symbols}
\alias{meetei_mayek}
\alias{meetei_mayek_extensions}
\alias{mende_kikakui}
\alias{meroitic_cursive}
\alias{meroitic_hieroglyphs}
\alias{miao}
\alias{miscellaneous_mathematical_symbols_a}
\alias{miscellaneous_mathematical_symbols_b}
\alias{miscellaneous_symbols_and_pictographs}
\alias{miscellaneous_technical}
\alias{modi}
\alias{modifier_tone_letters}
\alias{mongolian}
\alias{mro}
\alias{musical_symbols}
\alias{myanmar}
\alias{myanmar_extended_a}
\alias{myanmar_extended_b}
\alias{nabataean}
\alias{new_tai_lue}
\alias{nko}
\alias{number_forms}
\alias{ogham}
\alias{ol_chiki}
\alias{old_italic}
\alias{old_north_arabian}
\alias{old_permic}
\alias{old_persian}
\alias{old_south_arabian}
\alias{old_turkic}
\alias{optical_character_recognition}
\alias{oriya}
\alias{ornamental_dingbats}
\alias{osmanya}
\alias{pahawh_hmong}
\alias{pahlavi_inscriptional}
\alias{pahlavi_psalter}
\alias{palmyrene}
\alias{pau_cin_hau}
\alias{phags_pa}
\alias{phaistos_disc}
\alias{phoenician}
\alias{phonetic_extensions}
\alias{phonetic_extensions_supplement}
\alias{playing_cards}
\alias{private_use_area}
\alias{rejang}
\alias{rumi_numeral_symbols}
\alias{runic}
\alias{samaritan}
\alias{saurashtra}
\alias{sharada}
\alias{shavian}
\alias{shorthand_format_controls}
\alias{siddham}
\alias{sinhala}
\alias{sinhala_archaic_numbers}
\alias{small_form_variants}
\alias{sora_sompeng}
\alias{spacing_modifier_letters}
\alias{specials}
\alias{sundanese}
\alias{sundanese_supplement}
\alias{superscripts_and_subscripts}
\alias{supplemental_arrows_a}
\alias{supplemental_mathematical_operators}
\alias{supplemental_punctuation}
\alias{supplementary_private_use_area_a}
\alias{supplementary_private_use_area_b}
\alias{syloti_nagri}
\alias{syriac}
\alias{tagalog}
\alias{tagbanwa}
\alias{tags}
\alias{tai_le}
\alias{tai_tham}
\alias{tai_viet}
\alias{tai_xuan_jing_symbols}
\alias{takri}
\alias{tamil}
\alias{telugu}
\alias{thaana}
\alias{thai}
\alias{tibetan}
\alias{tifinagh}
\alias{tirhuta}
\alias{transport_and_map_symbols}
\alias{ugaritic}
\alias{unified_canadian_aboriginal_syllabics}
\alias{unified_canadian_aboriginal_syllabics_extended}
\alias{vai}
\alias{variation_selectors}
\alias{variation_selectors_supplement}
\alias{vedic_extensions}
\alias{vertical_forms}
\alias{warang_citi}
\alias{yi_radicals}
\alias{yi_syllables}
\alias{yijing_hexagram_symbols}
\alias{yijing_mono_di_and_trigrams}
\title{Unicode classes}
\description{
See \code{\link[rebus.unicode]{Unicode}}.
}
